feat: Add support for firmware reset command
Reset command has been implemented based on the previous repository
database file. Because of that, PV naming here is being kept, even
though it is not following the pattern used at LNLS of having the `-Cmd`
prefix for commands. It shall be updated in a future patch, which
changes all PV names at once.

Related the renaming aforementioned, the read-back record has not been
included herein, as it does not follow the standard but it is not really
used.

As this record does not really have a binary state, the label for its
zero value has been renamed to `reset`. On the other hand, `ONAM` has
been removed because it is not used elsewhere.