fix: Clip heater voltage to [0, 3.3]V interval
In firmware implementation [1], the tension limit for the heater is 3.3
Volts. Sending a value higher than that without configuring a clip
operation in EPICS leads the set-point request to be simply ignored by
the hardware. This leads to an inconsistency in the set-point and
read-back records, without any alert. By configuring a DRVH field, the
value is clipped to 3.3V whenever it's higher than that.

Since the heater expects to operate with a positive voltage, negative values
shall not be passed from the IOC to the device. Clipping to zero is also
performed in this case.

[1]: https://github.com/lnls-dig/rffe-uc-fw/blob/21329d8f9be430d373375f76a7168d4c50047cb2/rffe-app/scpi_rffe_cmd.c#L428