feat: Add heater controller support
Heater support include the switching between automatic and manual
heating control, besides the parameters for both methods. For the manual
heating setting, the voltage is used to indirectly control the
temperature. On the other hand, automatic control (through PID method)
is allows to specify a setpoint temperature to be reached and Kp, Ti,
and Td constants from both boards.

Periodic scan is used in read-back records so that PV has the updated
value when some external access is done to the device to update its
configuration. Moreover, when a value is updated, the corresponding
read-back value must be updated so that scripts can directly read back
the correct value. Therefore, it has also been used a forward link to
read-back from setpoint records. Note that it is used the `PROC` field
in the forwarding. This is because a non-passive scan is used in the
read-back records.

A description has been added to the `dfanout` records so that it is
clearer that the record is a shortcut for setting the same value for
both AC and BD boards.