feat: Add device temperature monitoring support
Monitor parameters ADEL and MDEL have been copied from the original IOC.
However, they may be not longer needed to be configured as they are.
This is because there is no forward link to any of them, which would be
the special situation in which an event would be triggered for `MDEL`
(`ADEL`) set to -1.