feat: Add firmware version record
The current version command from the firmware outputs four values: Apps,
NuttX OS, and firmware's Git commit hash along with the Git tag.

Therefore, the string length has been defined considering the following
points:

3 * 40: three Git commit hashes, 40 characters each
  +  3: commas for separating hashes and version tag
  +  3: three digits for version tag x.y.z
  +  2: two dots from the above number separation
  +  4: -rcX release candidate token (single digit)
  +  8: padding for more digits in any of the previous numbers.

The chosen padding length is so that the resulting number is the next
multiple of 10. This gives us room for more digits in each of the other
numbers, as well as extra tokens in version tags.